new_idris/backed/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9
gunicorn
python-dotenv
celery>=5.3
redis>=4.5
stripe>=5.0
django-cors-headers>=3.13.0
pytest-django>=4.5
